+++ tests/test_TimeseriesExtractor.py
@@ -16,7 +16,8 @@ def test_TimeseriesExtractor_no_parallel():
     bids_dir = os.path.join(dir, "ds000031_R1.0.4_ses001-022/ds000031_R1.0.4")
 
     pipeline_name = "fmriprep_1.0.0/fmriprep"
-    extractor.get_bold(bids_dir=bids_dir, session='002', runs="001",task="rest", pipeline_name=pipeline_name, tr=1.2, n_cores=1)
+    extractor.get_bold(bids_dir=bids_dir, session='002', runs="001",task="rest", pipeline_name=pipeline_name,
+                       tr=1.2, n_cores=1)
     
     print(extractor.subject_timeseries, flush=True)
 
@@ -71,6 +72,15 @@ def test_TimeseriesExtractor_no_parallel_no_session_w_custom_and_censoring():
     assert extractor.subject_timeseries["01"]["run-001"].shape[-1] == 426
     assert extractor.subject_timeseries["01"]["run-001"].shape[0] == 39
 
+    extractor = TimeseriesExtractor(parcel_approach=parcel_approach, standardize="zscore_sample",
+                                    use_confounds=True, detrend=True, low_pass=0.15, high_pass=0.01,
+                                    confound_names=confounds, fd_threshold={"threshold": 0.35,
+                                                                            "outlier_percentage": 0.0001})
+    
+    extractor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
+    
+    assert extractor.subject_timeseries == {}
+    
 def test_TimeseriesExtractor_dummy():
     dir = os.path.dirname(__file__)
 
@@ -106,5 +116,8 @@ def test_TimeseriesExtractor_check_exclusion():
     bids_dir = os.path.join(dir, "ds000031_R1.0.4_ses001-022/ds000031_R1.0.4")
 
     pipeline_name = "fmriprep_1.0.0/fmriprep"
-    extractor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, exclude_niftis=["sub-01_ses-002_task-rest_run-001_space-MNI152NLin2009cAsym_desc-preproc_bold.nii.gz"])
+    extractor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name,
+                       exclude_niftis=[
+                           "sub-01_ses-002_task-rest_run-001_space-MNI152NLin2009cAsym_desc-preproc_bold.nii.gz"
+                           ])
     assert len(extractor.subject_timeseries) == 0